+++ extensions/line/src/setup-surface.test.ts
@@ -6,8 +6,8 @@ import {
   resolveDefaultLineAccountId,
   resolveLineAccount,
 } from "../../../src/line/accounts.js";
-import { createRuntimeEnv } from "../../test-utils/runtime-env.js";
-import { createTestWizardPrompter, type WizardPrompter } from "../../test-utils/setup-wizard.js";
+import { createRuntimeEnv } from "../../../test/helpers/extensions/runtime-env.js";
+import { createTestWizardPrompter, type WizardPrompter } from "../../../test/helpers/extensions/setup-wizard.js";
 import { lineSetupAdapter, lineSetupWizard } from "./setup-surface.js";
 
 const lineConfigureAdapter = buildChannelSetupWizardAdapterFromSetupWizard({