+++ extensions/whatsapp/src/media.test.ts
@@ -7,8 +7,8 @@ import { resolveStateDir } from "../../../src/config/paths.js";
 import { resolvePreferredOpenClawTmpDir } from "../../../src/infra/tmp-openclaw-dir.js";
 import { optimizeImageToPng } from "../../../src/media/image-ops.js";
 import { mockPinnedHostnameResolution } from "../../../src/test-helpers/ssrf.js";
+import { captureEnv } from "../../../test/helpers/extensions/env.js";
 import { sendVoiceMessageDiscord } from "../../discord/src/send.js";
-import { captureEnv } from "../../test-utils/env.js";
 import {
   LocalMediaAccessError,
   loadWebMedia,